+++ CHANGELOG.md
@@ -5,6 +5,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [x.x.x] - xxxx-xx-xx
+
+### Added
+- **LowMem Time Effects:** Added `DelayLowMem` and `ReverbLowMem` with low memory usage, these prioritize low memory usage over high quality and performance.
+
+### Changed
+- **Delay:** Improved the quality and performance of `Delay`, it now supports sample accurate modulation as well.
+
 ## [0.7.0] - 2026-01-04
 
 ### Added